+++ doc/changelog.rst
@@ -2,6 +2,8 @@
 What's New
 ==========
 
+* Added `--preserve-permissions`.
+
 
 0.1008 (2015-02-26)
 ===================